+++ src/components/Settings/Generic.tsx
@@ -1,30 +1,47 @@
-import { Box } from "@mui/system";
 import { Setting } from "../../hooks/useSettings";
 import { MuiColorInput } from "mui-color-input";
+import Grid from "@mui/material/Grid";
 
 const Generic = ({
   setting,
   setSetting,
-  setThemePreview,
+  dispatchTheme,
 }: {
   setting: Setting | null;
   setSetting: CallableFunction;
-  setThemePreview: CallableFunction;
+  dispatchTheme: CallableFunction;
 }) => {
   const onChangeTheme = (color: string) => {
     setSetting({ ...setting, theme: color });
-    setThemePreview(color);
+    dispatchTheme({ type: "previewPrimary", primary: color });
   };
+
+  const onChangeThemeSecondary = (color: string) => {
+    setSetting({ ...setting, theme_secondary: color });
+    dispatchTheme({ type: "previewSecondary", secondary: color });
+  };
+
   return (
-    <Box>
-      <MuiColorInput
-        format="hex"
-        value={setting?.theme || "#1976d2"}
-        label="Theme"
-        fullWidth
-        onChange={onChangeTheme}
-      />
-    </Box>
+    <Grid container spacing={2}>
+      <Grid item md={6}>
+        <MuiColorInput
+          format="hex"
+          value={setting?.theme || "#1976d2"}
+          label="Primary"
+          fullWidth
+          onChange={onChangeTheme}
+        />
+      </Grid>
+      <Grid item md={6}>
+        <MuiColorInput
+          format="hex"
+          value={setting?.theme_secondary || "#ce93d8"}
+          label="Secondary"
+          fullWidth
+          onChange={onChangeThemeSecondary}
+        />
+      </Grid>
+    </Grid>
   );
 };