+++ src/styles/theme.ts
@@ -1,15 +1,15 @@
 import { createTheme } from "@mui/material/styles";
+import { Theme } from "../models";
 
-const appTheme = (
-  darkMode: boolean,
-  theme: string,
-  themePreview: string | null,
-) =>
+const appTheme = (darkMode: boolean, theme: Theme) =>
   createTheme({
     palette: {
       mode: darkMode ? "dark" : "light",
       primary: {
-        main: themePreview ? themePreview : theme,
+        main: theme.primaryPreview || theme.primary,
+      },
+      secondary: {
+        main: theme.secondaryPreview || theme.secondary,
       },
     },
   });